Allow styles to depend on Assets
This is a necessity to allow doodads to incorporate assets like
background images into their css.

I've tried to make the change as backwards-compatible as possible,
leaving most of the existing funtions alone. The `.css` property on the
returned Styles instance still needs those asset paths though, so this
is still a major update.