Fix attribute translation
I was taking the elm names as the source of truth, but elm is
camelCased, and attributes use `-`. I did some quick sublimefu which in
hindsight I maybe should have grep-fu'ed so it'd be reproducible, but I
got a mapping of elm->svg->elm.

This was borking the translation of the calendar icon, specifically.  It
was using `fill-rule`.