Updated Assets
They're looking OK'ish, the stuff that used to be tiny in PNG is taking
up all the container space in SVG, which makes me wonder if we'll have
to do CSS workarounds for sizing.

Also, I'm not deleting styles which I maybe should, but there's icons we
distinguish by color like hint and bulb.