Lighter Icons API
If we're embedding assets I think we can do away with the extensible
records stuff and trust whatever users pass in to the `custom*`
functions.

This might make usage of the icons a little more verbose (it's evident
by trying to change the Example/Icons.elm), but I'm getting the feeling
we shouldn't be adding images here that we don't host.