Options for SegmentedControl's width: FitContent | FillContainer
Until V4 the control would always fit its contents. On V5 that behaviour was
changed to always fill the container. This means that to limit the element's
width one had to limit the value of its container (which might be hard to do
when the labels of the tabs are unknown or might change).

We now support both settings, and the user has to decide what to do for each use
case.