Allow V2 and V3 buttons to grow vertically.
Presently, buttons are being displayed to students in quizzes with more text
than a button will show. The overflow is not visible and it's making it hard for
these students to use the site. A button should always show the text that it
contains, even if it causes layout weirdness.