External tracking should work with ctrl/cmd- clicks
A button with `linkExternalWithTracking` was previously using
`EventExtras.onClickForLinkWithHref`, which was causing the event
handler to not be called when the user clicks on the link while
pressing cmd or ctrl. This is actually fine for tacking purposes,
as the page should still be open in the old tab, allowing the tracking
request to complete.

It's not clear why this particular event helper was used here for
the `ExternalWithTracking` case, but we can't find a downside for
swapping it out for a regular `Events.onClick`.