add a separate checked attribute for writing tests against
The basic problem here: `node.checked = true` in a real browser does not
modify the DOM; it only modifies the internal state of the element. It shows
up in our tests because `Test.Html.Query` makes all the elements and properties
visible for testing.

To get around this, we just add a `data-nri-checked` for automated tests
to use.