wrap grey block with a button, not a div
This is so we can select the radio button when clicking anywhere in the
grey. So why not a <label>? Then the full contents of the disclosure
would be read out as the label content when picking options. That
content *describes* the radio button but it does not *label* it, and
should not be read out before "selected, n of m" and the rest of the
radio helper voiceover content.