Adds another utility for creating safe Id strings
While implementing something with Checkboxv7 we noticed an issue where
when the id was being created from the label it would break the filter
on the svg causing the fill to be black. In this case it was due to the
label having something like "Angela's account" in it which broke the id
on the comma.